fix: rename api to services
